Add Facia signposting override for editorally curated index pages
