+++ facia/app/model/FaciaPage.scala
@@ -1,6 +1,6 @@
 package model
 
-import common.Edition
+import common.{NavItem, Edition}
 import dfp.DfpAgent
 import play.api.libs.json.{JsString, JsValue}
 
@@ -33,6 +33,8 @@ case class FaciaPage(id: String,
   override def hasPageSkin(edition: Edition) = DfpAgent.isPageSkinned(adUnitSuffix, edition)
 
   def allItems = collections.map(_._2).flatMap(_.items).distinct
+
+  override def customSignPosting: Option[NavItem] = FaciaSignpostingOverrides(id)
 }
 
 object FaciaPage {
